+++ README.md
@@ -73,17 +73,3 @@ for structure in sample_structures:
     mofchecker = MOFChecker.from_cif(structure)
     results.append(mofchecker.get_mof_descriptors())
 ```
-
-### CLI
-
-For example, you can use
-
-```bash
-mofchecker <cif> --has-oms
-```
-
-You can get an overview over all options with
-
-```bash
-mofchecker --help
-```
