+++ mofchecker/checks/local_structure/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 from .overcoordinated_carbon import OverCoordinatedCarbonCheck
+from .overcoordinated_hydrogen import OverCoordinatedHydrogenCheck
 from .overlapping_atoms import AtomicOverlapCheck
 from .undercoordinated_carbon import UnderCoordinatedCarbonCheck
-from .overcoordinated_hydrogen import OverCoordinatedHydrogenCheck
+from .overcoordinated_nitrogen import OverCoordinatedNitrogenCheck
+from .undercoordinated_nitrogen import UnderCoordinatedNitrogenCheck
